+++ worker/src/RTC/RTCP/XrReceiverReferenceTime.cpp
@@ -35,7 +35,7 @@ namespace RTC
 			MS_TRACE();
 
 			MS_DUMP("<ReceiverReferenceTime>");
-			MS_DUMP("  block type : %" PRIu8, this->type);
+			MS_DUMP("  block type : %" PRIu8, (uint8_t)this->type);
 			MS_DUMP("  reserved   : 0");
 			MS_DUMP("  length     : 2");
 			MS_DUMP("  ntp sec    : %" PRIu32, GetNtpSec());
